Add InlineSign API
