+++ sop-java/src/main/java/sop/operation/InlineSign.java
@@ -4,6 +4,22 @@
 
 package sop.operation;
 
-public interface InlineSign {
+import sop.enums.InlineSignAs;
+import sop.exception.SOPGPException;
+
+import java.io.InputStream;
+
+public interface InlineSign extends AbstractSign<InlineSign> {
+
+    /**
+     * Sets the signature mode.
+     * Note: This method has to be called before {@link #key(InputStream)} is called.
+     *
+     * @param mode signature mode
+     * @return builder instance
+     *
+     * @throws sop.exception.SOPGPException.UnsupportedOption if this option is not supported
+     */
+    Sign mode(InlineSignAs mode) throws SOPGPException.UnsupportedOption;
 
 }
