+++ sop-java/src/main/java/sop/operation/Sign.java
@@ -4,24 +4,12 @@
 
 package sop.operation;
 
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.nio.charset.Charset;
-
-import sop.ReadyWithResult;
-import sop.SigningResult;
 import sop.enums.SignAs;
 import sop.exception.SOPGPException;
 
-public interface Sign {
+import java.io.InputStream;
 
-    /**
-     * Disable ASCII armor encoding.
-     *
-     * @return builder instance
-     */
-    Sign noArmor();
+public interface Sign extends AbstractSign<Sign> {
 
     /**
      * Sets the signature mode.
@@ -34,71 +22,4 @@ public interface Sign {
      */
     Sign mode(SignAs mode) throws SOPGPException.UnsupportedOption;
 
-    /**
-     * Add one or more signing keys.
-     *
-     * @param key input stream containing encoded keys
-     * @return builder instance
-     *
-     * @throws sop.exception.SOPGPException.KeyIsProtected if the key is password protected
-     * @throws sop.exception.SOPGPException.BadData if the {@link InputStream} does not contain an OpenPGP key
-     * @throws IOException in case of an IO error
-     */
-    Sign key(InputStream key) throws SOPGPException.KeyIsProtected, SOPGPException.BadData, IOException;
-
-    /**
-     * Add one or more signing keys.
-     *
-     * @param key byte array containing encoded keys
-     * @return builder instance
-     *
-     * @throws sop.exception.SOPGPException.KeyIsProtected if the key is password protected
-     * @throws sop.exception.SOPGPException.BadData if the byte array does not contain an OpenPGP key
-     * @throws IOException in case of an IO error
-     */
-    default Sign key(byte[] key) throws SOPGPException.KeyIsProtected, SOPGPException.BadData, IOException {
-        return key(new ByteArrayInputStream(key));
-    }
-
-    /**
-     * Provide the decryption password for the secret key.
-     *
-     * @param password password
-     * @return builder instance
-     */
-    default Sign withKeyPassword(String password) {
-        return withKeyPassword(password.getBytes(Charset.forName("UTF8")));
-    }
-
-    /**
-     * Provide the decryption password for the secret key.
-     *
-     * @param password password
-     * @return builder instance
-     */
-    Sign withKeyPassword(byte[] password);
-
-    /**
-     * Signs data.
-     *
-     * @param data input stream containing data
-     * @return ready
-     *
-     * @throws IOException in case of an IO error
-     * @throws sop.exception.SOPGPException.ExpectedText if text data was expected, but binary data was encountered
-     */
-    ReadyWithResult<SigningResult> data(InputStream data) throws IOException, SOPGPException.ExpectedText;
-
-    /**
-     * Signs data.
-     *
-     * @param data byte array containing data
-     * @return ready
-     *
-     * @throws IOException in case of an IO error
-     * @throws sop.exception.SOPGPException.ExpectedText if text data was expected, but binary data was encountered
-     */
-    default ReadyWithResult<SigningResult> data(byte[] data) throws IOException, SOPGPException.ExpectedText {
-        return data(new ByteArrayInputStream(data));
-    }
 }
